Day 4
Today I just added a image to the home page and made a media query to make it responsive. After that I pondered over a template for my three main pages. Decided on one then cleaned up my code to be less redundant and hectic. Tomorrow I will add the leftover images to the site, create something for the home page , and create the about page.